requirements.txt
Flask==2.3.3
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
google-search-results==2.4.2
selenium==4.20.0
webdriver-manager==4.0.1
